fix deploying issues
